reorg

+++ app/model/Msg.scala
@@ -1,4 +1,4 @@
 package model
 
-case class Loc(address : String, lng : BigDecimal, lat : BigDecimal)
+case class Loc(address : String, lng : Float, lat : Float)
 case class Msg(name : String, location : Loc, title : String, body : Option[String])
